Move subject-page term API to edge runtime
Cuts cold function-start on first-ever term switch from ~2s (Node
serverless) to ~500ms (edge). The route is called whenever a user
toggles terms on the subject page, so cold starts are common.

Scope grew beyond a one-line `runtime = "edge"` because the route's
transitive deps weren't all edge-safe:

- `lib/institutions.ts` used `fs.readFileSync` + `process.cwd()` to
  load per-state institution JSONs. Switched to static JSON imports
  (~400 KB across all 15 states, well under edge bundle limits).
- `lib/states/registry.ts` used `require()` to lazy-load state
  configs. Replaced with static imports.
- `lib/courses.ts` imported `./geo` (zip-code lookup, uses `fs`) at
  module top level for the cross-college search helper. Extracted
  the search helper to `lib/courses-search.ts` so the Supabase core
  stays fs-free. Only consumer (`/api/[state]/courses/search`)
  updated to the new path.
- `lib/transfer.ts` still has a 19-51 MB fs fallback and can't be
  bundled into edge. Inlined a targeted Supabase `.or()` query in
  the route to build the per-request transfer lookup directly —
  ~15 lines of duplication vs. dragging the whole module in.

Verified: `tsc --noEmit` clean, `next build` emits no
Node-API-in-edge warnings, dev preview term switch works, returns
82 courses + 10 transfer-lookup keys for the test case.
